+++ src/services/BoardService.js
@@ -34,13 +34,34 @@ async function getAll( projectId ) {
  */
 async function deleteBoard( id ) {
     try {
-        await Board.deleteOne( { _id: id } ).exec();
-        return true;
+        if ( id !== undefined ) {
+            await Board.deleteOne( { _id: id } ).exec();
+            return true;
+        }
+        return false;
     } catch ( e ) {
         return false;
     }
 }
 
+/**
+ * Update the board with the given values.
+ * @param {String} id, the id of the board.
+ * @param {*} data, the data for updating the board.
+ * @param {String[]} columns, the columns to add to the board.
+ */
+async function updateBoard( id, data, columns ) {
+    let parsedData = data;
+    if ( columns !== undefined && columns.length > 0 ) {
+        parsedData = {
+            ...data,
+            $push: { columns: { $each: columns } },
+        };
+    }
+    const board = await Board.findOneAndUpdate( { _id: id }, parsedData, { new: true } ).exec();
+    return board;
+}
+
 module.exports = {
-    createBoard, getAll, deleteBoard,
+    createBoard, getAll, deleteBoard, updateBoard,
 };
